src/system: use /etc/machine-id as machine identifier
At least on my raspberrypi 3, with a mainline kernel, /proc/cpuinfo does
not contain a `Serial:` line.

It neither does on my other system, so it was impossible to run the
application in release mode.

This now instead reads from `/etc/machine-id`, as per
https://www.freedesktop.org/software/systemd/man/latest/machine-id.html

That one is guaranteed to be stable across reboots, as long as we don't
loose the root filesystem, and works on pretty much any Linux system
out there.

It's also possible to influence its generation, as described in
https://www.freedesktop.org/software/systemd/man/latest/machine-id.html#Initialization

This would allow passing some unique identifier determined in u-boot
to be passed to the kernel cmdline, for example, or a early-boot systemd
unit file that interacts with
https://www.freedesktop.org/software/systemd/man/latest/systemd-machine-id-setup.html#

Point being we don't need to make this too hardware-specific.

I kept the `id` thing in the config around for now, so when testing
on MacOS, it can be provided in the config to get a stable identifier
there.