mqtt: handle display demuxing
This updates the mqtt listener to maintain a single connection to a
broker, but allow different displays to be registered, and a tx to write
into to be passed for each individual one.

We currently still only use it for a single display, but it makes a lot
of things much easier.